+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### 0.2.6
 
 * Fix default formatter not defined error.
+* Add ellipsis to options that don't fit into the screen.
+* Re-render on terminal resize.
 
 ### 0.2.5